Added stream interfaces, tweaks - Added Stream interfaces - Added converters between IO.Stream and Stream interfaces - Fixed return type of stream ReadUInt16 extension method - Switched to assert-on-failure for contracts

+++ Streams/StreamExtensions.vb
@@ -104,7 +104,7 @@ Namespace Streams
         ''' </summary>
         <Extension()>
         Public Function ReadUInt16(ByVal stream As IO.Stream,
-                                   Optional ByVal byteOrder As ByteOrder = ByteOrder.LittleEndian) As UInt32
+                                   Optional ByVal byteOrder As ByteOrder = ByteOrder.LittleEndian) As UInt16
             Contract.Requires(stream IsNot Nothing)
             Contract.Requires(stream.CanRead)
             Return stream.ReadBytesExact(length:=2).ToUInt16(byteOrder)
